split kmachine
