+++ game/kernel/jak2/kmachine.h
@@ -1,5 +1,11 @@
 #pragma once
 
 namespace jak2 {
+void InitParms(int argc, const char* const* argv);
+void InitIOP();
+int InitMachine();
 
+struct MouseInfo {
+
+};
 }
