+++ game/kernel/kscheme.cpp
@@ -11,7 +11,6 @@
 #include "kdgo.h"
 #include "klink.h"
 #include "klisten.h"
-#include "kmachine.h"
 #include "kmemcard.h"
 
 #include "common/common_types.h"
@@ -31,6 +30,7 @@
 #include "game/kernel/common/kscheme.h"
 #include "game/mips2c/mips2c_table.h"
 #include "game/kernel/jak1/kprint.h"
+#include "game/kernel/jak1/kmachine.h"
 
 using namespace jak1_symbols;
 
@@ -1111,52 +1111,6 @@ u64 method_set(u32 type_, u32 method_id, u32 method) {
   return method;
 }
 
-extern "C" {
-// defined in asm_funcs.asm
-#ifdef __linux__
-uint64_t _call_goal_asm_linux(u64 a0, u64 a1, u64 a2, void* fptr, void* st_ptr, void* offset);
-uint64_t _call_goal_on_stack_asm_linux(u64 rsp,
-                                       u64 u0,
-                                       u64 u1,
-                                       void* fptr,
-                                       void* st_ptr,
-                                       void* offset);
-#elif _WIN32
-uint64_t _call_goal_asm_win32(u64 a0, u64 a1, u64 a2, void* fptr, void* st_ptr, void* offset);
-uint64_t _call_goal_on_stack_asm_win32(u64 rsp, void* fptr, void* st_ptr, void* offset);
-#endif
-}
-
-/*!
- * Wrapper around _call_goal_asm for calling a GOAL function from C.
- * Calls from the parent stack.
- */
-u64 call_goal(Ptr<Function> f, u64 a, u64 b, u64 c, u64 st, void* offset) {
-  // auto st_ptr = (void*)((uint8_t*)(offset) + st); updated for the new compiler!
-  void* st_ptr = (void*)st;
-
-  void* fptr = f.c();
-#ifdef __linux__
-  return _call_goal_asm_linux(a, b, c, fptr, st_ptr, offset);
-#elif _WIN32
-  return _call_goal_asm_win32(a, b, c, fptr, st_ptr, offset);
-#endif
-}
-
-/*!
- * Wrapper around _call_goal_asm_on_stack for switching stacks and calling a GOAL function there.
- */
-u64 call_goal_on_stack(Ptr<Function> f, u64 rsp, u64 st, void* offset) {
-  void* st_ptr = (void*)st;
-
-  void* fptr = f.c();
-#ifdef __linux__
-  return _call_goal_on_stack_asm_linux(rsp, 0, 0, fptr, st_ptr, offset);
-#elif _WIN32
-  return _call_goal_on_stack_asm_win32(rsp, fptr, st_ptr, offset);
-#endif
-}
-
 /*!
  * Call a GOAL method of a given type.
  */
@@ -2038,7 +1992,7 @@ s32 InitHeapAndSymbol() {
   InitListener();
 
   // Do final initialization, including loading and initializing the engine.
-  InitMachineScheme();
+  jak1::InitMachineScheme();
 
   // testing stuff:
   make_function_symbol_from_c("test-function", (void*)test_function);
